// src/component/admin-features/request-details/types/entityProfileApprovalTypes.ts
export interface EntityProfilePendingApprovalItem {
  workflowId: string;
  workflowType?: string;
  fiCode: string;
  institutionName?: string;
  status: string;
  submittedOn: string;
  submittedBy: string;
  acknowledgementNo: string;
  reId?: string;
  userId?: string;
  createdAt?: string;
  updatedAt?: string;
  noOfApprovalRequired?: number;
}